requirements.txt
schedule>=1.1.0
stomper>=0.4.3
websocket-client>=1.2.1
uuid>=1.3